step1の最後のbabelによるjsxトランスパイルがうまく行かないところまで
